organized address component on other folder inside user area
