+++ src/pages/UserArea/index.tsx
@@ -8,17 +8,13 @@ import {
   Drawer,
   Cart,
   CartOrderCard,
-  AddressCard,
   TextInput,
   ButtonMain,
-  CheckBox,
   FormWrapper,
 } from '../../components';
 import { GlobalUserContext } from '../../contexts';
 import {
   ADDRESS_EDITING_STATUS,
-  ADDRESS_EDITING_STATUS_MESSAGE,
-  ADDRESS_EDITING_STATUS_MESSAGE_BUTTON,
   DRAWER_VALUES,
   USER_AREA,
   ZipCodeMask,
@@ -36,6 +32,7 @@ import {
   isValidEmail,
   getBackFormattedDate,
 } from '../../utils';
+import AddressesComponent from './Address';
 import {
   initialAddressFormErrors,
   initialAddressFormValues,
@@ -44,17 +41,9 @@ import {
 import {
   AccountInformationSubtitle,
   AccountInformationWrapper,
-  AddressCardsWrapper,
-  AddressesWrapper,
-  DeleteWrapper,
   DrawerSelectedWrapper,
   DrawerWrapper,
-  FormHeader,
   InternWrapper,
-  IoMdClose,
-  NewAddressFormWrapper,
-  NewAddressInternFormWrapper,
-  NewAddressTitle,
   Wrapper,
 } from './styles';
 
@@ -357,133 +346,23 @@ const UserArea: React.FC = () => {
   };
 
   const addressesComponent = () => (
-    <AddressesWrapper id={USER_AREA.ADDRESSES_WRAPPER}>
-      <NewAddressFormWrapper openState={showForm}>
-        <FormWrapper>
-          <FormHeader>
-            <NewAddressTitle>
-              {ADDRESS_EDITING_STATUS_MESSAGE[formEditingState] ||
-                'Novo endereço'}
-            </NewAddressTitle>
-            <DeleteWrapper>
-              <IoMdClose size={30} onClick={handleShowForm} />
-            </DeleteWrapper>
-          </FormHeader>
-          <NewAddressInternFormWrapper>
-            <TextInput
-              label="Nome do endereço"
-              inputProps={{
-                placeholder: 'Casa da vó',
-                value: addressFormValues.name,
-                onChange: ({ target }) =>
-                  genericAddressFormChange(target, 'name'),
-              }}
-              error={formErrors.name}
-              fullWidth
-            />
-            <TextInput
-              label="CEP"
-              inputProps={{
-                placeholder: '123456-000',
-                inputMode: 'numeric',
-                value: addressFormValues.cep,
-                maxLength: 9,
-                onChange: ({ target }) => handleZipCodeChange(target),
-              }}
-              error={formErrors.cep}
-              fullWidth
-            />
-            <TextInput
-              label="Rua"
-              inputProps={{
-                placeholder: 'Alfredo Neves',
-                value: addressFormValues.street,
-                onChange: ({ target }) =>
-                  genericAddressFormChange(target, 'street'),
-              }}
-              error={formErrors.street}
-              fullWidth
-            />
-            <TextInput
-              label="Número"
-              inputProps={{
-                placeholder: '1234',
-                inputMode: 'text',
-                maxLength: 20,
-                value: addressFormValues.number,
-                onChange: ({ target }) =>
-                  genericAddressFormChange(target, 'number'),
-              }}
-              error={formErrors.number}
-              fullWidth
-            />
-            <TextInput
-              label="Bairro"
-              inputProps={{
-                placeholder: 'Jardim das Palmeiras',
-                value: addressFormValues.district,
-                maxLength: 80,
-                onChange: ({ target }) =>
-                  genericAddressFormChange(target, 'district'),
-              }}
-              error={formErrors.district}
-              fullWidth
-            />
-            <TextInput
-              impossibleToEdit={isValidZipCode}
-              label="Estado"
-              inputProps={{
-                placeholder: 'RS',
-                contentEditable: false,
-                value: addressFormValues.uf,
-                onChange: ({ target }) =>
-                  genericAddressFormChange(target, 'uf'),
-              }}
-              error={formErrors.uf}
-              fullWidth
-            />
-            <TextInput
-              impossibleToEdit={isValidZipCode}
-              label="Cidade"
-              inputProps={{
-                placeholder: 'Alegrete',
-                value: addressFormValues.city,
-                onChange: ({ target }) =>
-                  genericAddressFormChange(target, 'city'),
-              }}
-              error={formErrors.city}
-              fullWidth
-            />
-            <CheckBox
-              label="Endereço padrão?"
-              checkBoxProps={{
-                checked: addressFormValues.defaultAddress,
-                onClick: onCheckboxClick,
-              }}
-            />
-          </NewAddressInternFormWrapper>
-          <ButtonMain
-            onClick={ADDRESS_EDITING_STATUS_FUNCTION[formEditingState]}
-          >
-            {ADDRESS_EDITING_STATUS_MESSAGE_BUTTON[formEditingState]}
-          </ButtonMain>
-        </FormWrapper>
-      </NewAddressFormWrapper>
-      <AddressCardsWrapper>
-        <AddressCard key={0} isNew onNewPress={onNewPress} />
-        {user?.adresses.map(address => (
-          <AddressCard
-            address={address}
-            key={address.id || 0}
-            onEditPress={onEditPress}
-            onDeletePress={() => handleDeleteAddress(address.id)}
-            onDefineDefaultPress={() =>
-              onDefineDefaultPress(address, address.id)
-            }
-          />
-        ))}
-      </AddressCardsWrapper>
-    </AddressesWrapper>
+    <AddressesComponent
+      ADDRESS_EDITING_STATUS_FUNCTION={ADDRESS_EDITING_STATUS_FUNCTION}
+      addressFormValues={addressFormValues}
+      formEditingState={formEditingState}
+      formErrors={formErrors}
+      genericAddressFormChange={genericAddressFormChange}
+      handleDeleteAddress={handleDeleteAddress}
+      handleShowForm={handleShowForm}
+      handleZipCodeChange={handleZipCodeChange}
+      isValidZipCode={isValidZipCode}
+      onCheckboxClick={onCheckboxClick}
+      onDefineDefaultPress={onDefineDefaultPress}
+      onEditPress={onEditPress}
+      onNewPress={onNewPress}
+      showForm={showForm}
+      user={user}
+    />
   );
 
   const accountInformationComponent = () => (
